refactor(rust): extract contexts.rs with context snippet commands
